Fixed filterMoves function and moveKnight function
filterMoves was previously becoming too big
because it was editing a copy of the array and
reiterating over it after it made edits which
meant switching indexes around. I changed it to
push the values we want to a new array instead of
slicing them out of the copy of the array. This is
easier to comprehend and has less lines of code.

moveKnight function will now work with a start
position of 0,0 and end position of 3,3. It still
needs further work, especially on the positions
visited variable and how this is handled.